More restart file modifications
Add global strings for yapioRestartFileName and yapioMkstempSuffix.

The yapioMkstempSuffix is MPI_Bcast now instead of the tail of
yapioTestFileName.  yapioMkstempSuffix is now used in snprintf for the
restart file.

The restart file is now unlinked if keep-file is false.

Yapio suffix is now applied to the output so that users can see which
temp string was used to name the output file names:
    0.35 IzfmMM.00.00: swL---  48.33 MiB/s