Add stonewalling feature
This patch introduces stonewalling (-S <num seconds>) which can be used
to terminate a test after some number of seconds.

The implementation uses a thread which wakes up every second and checks
for the deadline.  When the deadline has been reached, this thread sets
a variable to 'true' and this variable is checked in the program's
main I/O loop.  When a stonewall has been hit, the program will not
fsync it's dirty contents.

When using -k, yapio will now write its metadata files after the last
write phase of the test (as opposed to after the last test).  This
change is to allow for metadata contents to be written before the
stonewall deadline has been reached.  If the deadline is reached before
the last write test has completed, and -k has been specified, the
program will invalidate -k and cause the output files to be removed.

Note that results from stonewalled tests attempt to provide the averaged
bandwidth by calculating the number of blocks that were processed before
the deadline expired.

Lastly, tests which have been stonewalled are marked with an 'X' in
their results display line.