Avoid using mpi group context-specific pointer
yapio_display_result() was accessing a yapio_test_group_t pointer
through yapio_test_ctx_t.  This was causing a segfault when the
reporting thread's mpi group was different from that of the reporter.

This patch disallows the use of ytc->ytc_group in yapio_display_result()
and replaces it with a function parameter.